+++ Thunderstore/CHANGELOG.md
@@ -1,4 +1,8 @@
-# v0.3.0
+# v0.3.1
+
+- Switched from using quaternion on entity replacements, I realised that if even I don't get how they work, I can't expect anyone else to also understand that.
+
+## v0.3.0
 
 - **Update to v73. This version and versions after are not compatible with v72 or lower! (revert to v0.2.16 for v72)**
 - Added support for `Moons` and `Dungeons`. Please note that this is very experimental at the moment (compatibility for LLL is also unknown)